pyproject.toml
[tool.poetry]
name = "tatoxer_gendiff"
version = "0.2.0"
description = ""
authors = ["Anton Selifanov <Tatoxer@gmail.com>"]
packages = [
  { include = "gendiff" },
]

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.dependencies]
python = "^3.8"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
pytest = "^5.4.1"
pytest-cov = "^2.8.1"
PyYAML = "^5.3.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
